Add files via upload
Voronoi tesselation - Get a Voronoi diagram and calculate the area and perimeter of the cells that have proper boundaries. Also we can find out the number of neighbours for each cell - the coordinates of the shared edge and its length.  This will be used in the next code where we calculate the energy function of all the cells and perform a Monte Carlo simulation for energy minimization.